refactor(voice): move PlaceVoiceTestCallAction into Actions/Voice
Address review: group voice actions under a dedicated Voice/ subdir. Updates
the namespace to Kanvas\Intelligence\Agents\Actions\Voice and the mutation's
use statement accordingly.
